Atualiza tipos de dados em funções de waypoint para double e ajusta chamadas correspondentes, melhorando a precisão das coordenadas geográficas.

+++ src/guidance/MissionManager.h
@@ -33,10 +33,10 @@ public:
     // 1. Controle da Missão
     void loadMission(); // (Mockup) Carrega WPs para a RAM
     void update(float current_lat, float current_lon, float current_alt, bool is_rth_active);
-    void saveWaypoint(uint8_t index, float lat, float lon, float alt, float speed);
+    void saveWaypoint(uint8_t index, double lat, double lon, float alt, float speed);
 
     // 2. Definir o Ponto de Retorno (Home)
-    void setHome(float lat, float lon, float alt);
+    void setHome(double lat, double lon, float alt);
 
     // 3. Getters para o L1 Guidance e TECS
     float getActiveLat();
